backend/embeddings/requirements.txt
flask
langchain-nvidia-ai-endpoints
python-dotenv
